+++ packages/browser/src/client.ts
@@ -124,7 +124,7 @@ export class BrowserClient extends BaseClient<BrowserOptions> {
   protected _setupTransport(): Transport {
     if (!this._options.dsn) {
       // We return the noop transport here in case there is no Dsn.
-      return super._setupTransport();
+      return undefined;
     }
 
     const transportOptions = {
